+++ web/src/lib.rs
@@ -349,7 +349,7 @@ pub async fn latest_block(network_param: Option<String>) -> String {
 		light_client::process_block(db, &network_client, 99.9, header, received_at, lc_sender)
 			.await
 			.unwrap()
-			.map(|confidence| format!("{confidence}"))
+			.map(|block_verified| format!("{:?}", block_verified.confidence))
 			.unwrap_or("none".to_string());
 
 	format!("Confidence for block {} is {confidence}.", block.number())